requirements.txt
academictorrents==2.3.3
albumentations==1.4.3
einops==0.7.0
Flask==3.0.3
Flask_Cors==4.0.0
matplotlib==3.8.4
more_itertools==10.2.0
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.9.0.80
Pillow==9.4.0
Pillow==10.3.0
pydicom==2.4.4
pytorch_fid==0.3.0
pytorch_lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scipy==1.13.0
segmentation_models_pytorch==0.3.3
skimage==0.0
torch==2.2.2
torchsummary==1.5.1
torchvision==0.17.2
tqdm==4.66.2
transformers==4.39.3
typing_extensions==4.11.0
